finishing ui goal view and add id argument for load data in goal screen

+++ composeApp/src/androidMain/kotlin/com/kakapo/oakane/presentation/feature/home/navigation/HomeNavigation.kt
@@ -17,7 +17,7 @@ fun NavGraphBuilder.homeScreen(
     navigateToAddTransaction: (Long) -> Unit,
     navigateToTransactions: () -> Unit,
     navigateToAddGoal: () -> Unit,
-    navigateToGoal: () -> Unit
+    navigateToGoal: (Long) -> Unit
 ) {
     composable(HOME_ROUTE) {
         HomeRoute(
